requirements.txt
accelerate==0.19.0
albumentations==1.3.0
alembic==1.11.1
appdirs==1.4.4
arrow==1.2.3
astunparse==1.6.3
audioread==3.0.0
blis==0.7.9
brotlipy==0.7.0
catalogue==2.0.8
certifi==2023.5.7
click==8.1.3
cmaes==0.9.1
colorednoise==2.1.0
colorlog==6.7.0
conda==23.3.1
conda-build==3.24.0
confection==0.0.4
contourpy==1.0.7
cycler==0.11.0
cymem==2.0.7
debugpy==1.6.7
diffusers==0.16.1
dnspython==2.3.0
docker-pycreds==0.4.0
exceptiongroup==1.1.1
expecttest==0.1.4
fonttools==4.39.4
fqdn==1.5.1
fsspec==2023.5.0
ftfy==6.1.1
greenlet==2.0.2
huggingface-hub==0.14.1
hypothesis==6.75.2
imageio==2.28.1
ipykernel==6.23.1
ipython-genutils==0.2.0
ipywidgets==8.0.6
isoduration==20.11.0
joblib==1.2.0
jsonpointer==2.1
jupyterlab-widgets==3.0.7
kaggle==1.5.13
kiwisolver==1.4.4
langcodes==3.3.0
lazy_loader==0.2
librosa==0.10.0.post2
lightgbm==3.3.5
llvmlite==0.40.0
Mako==1.2.4
matplotlib==3.7.1
mkl-fft==1.3.6
mkl-service==2.4.0
mpmath==1.3.0
msgpack==1.0.5
murmurhash==1.0.9
nbclassic==1.0.0
networkx==3.1
nnAudio==0.3.2
notebook==6.5.4
numba==0.57.0
opencv-python-headless==4.7.0.72
optuna==3.2.0
pandas==2.0.1
pathtools==0.1.2
pathy==0.10.1
Pillow==9.4.0
pooch==1.6.0
preshed==3.0.8
protobuf==4.23.2
pyarrow==12.0.0
pydantic==1.10.7
pyparsing==3.0.9
pyrsistent==0.19.3
python-etcd==0.4.5
python-slugify==8.0.1
pytorch-metric-learning==2.1.2
PyWavelets==1.4.1
qudida==0.0.4
regex==2023.5.5
safetensors==0.3.1
scikit-image==0.20.0
scikit-learn==1.2.2
scipy==1.10.1
sentry-sdk==1.25.0
setproctitle==1.3.2
smart-open==6.3.0
sortedcontainers==2.4.0
soundfile==0.12.1
soxr==0.3.5
spacy==3.5.3
spacy-legacy==3.0.12
spacy-loggers==1.0.4
SQLAlchemy==2.0.15
srsly==2.4.6
sympy==1.12
text-unidecode==1.3
thinc==8.1.10
threadpoolctl==3.1.0
tifffile==2023.4.12
timm==0.9.2
tokenizers==0.13.3
torch==2.0.1
torchaudio==2.0.2
torchelastic==0.2.2
torchlibrosa==0.1.0
torchtext==0.15.2
torchvision==0.15.2
tornado==6.3.2
transformers==4.29.1
triton==2.0.0
typer==0.7.0
types-dataclasses==0.6.6
tzdata==2023.3
uri-template==1.2.0
wandb==0.15.3
wasabi==1.1.1
webcolors==1.13
webencodings==0.5.1
widgetsnbextension==4.0.7